+++ Magic/CHANGELOG.md
@@ -7,6 +7,7 @@
      See: https://github.com/elBukkit/MagicPlugin/tree/master/src/main/resources/examples/rpg/attributes.yml
    - Add /mage attribute command for listing and setting attributes
  - Add "Magic.commands.mage.others" to allow someone to use /mage commands on other players
+ - Add /mage switch command to switch classes for a player (combo of lock, unlock and activate)
 
 ## 7.3.1